+++ js/src/patina/Patina.routes.tsx
@@ -1,12 +1,12 @@
 import { HomePage } from '@/patina/pages/home/Home.page.tsx'
 import { MissionPage } from '@/patina/pages/about/mission/Mission.page.tsx'
 import { TeamPage } from '@/patina/pages/about/team/Team.page.tsx'
-import { MentorshipPage as MentorshipFa24Page } from '@/patina/pages/mentorship-fa24/Mentorship.page.tsx'
-import { MentorshipPage as MentorshipSp25Page } from '@/patina/pages/mentorship-sp25/Mentorship.page.tsx'
+import { MentorshipPage as MentorshipFa24Page } from '@/patina/pages/past-programs/mentorship-fa24/Mentorship.page.tsx'
+import { MentorshipPage as MentorshipSp25Page } from '@/patina/pages/past-programs/mentorship-sp25/Mentorship.page.tsx'
 import { MentorshipPage as MentorshipFa25Page } from '@/patina/pages/mentorship-fa25/Mentorship.page.tsx'
 import { CommunityPage } from '@/patina/pages/community/Community.page.tsx'
-import { MentorPage as MentorFa24Page } from '@/patina/pages/mentor-fa24/Mentor.page.tsx'
-import { MentorPage as MentorSp25Page } from '@/patina/pages/mentor-sp25/Mentor.page.tsx'
+import { MentorPage as MentorFa24Page } from '@/patina/pages/past-programs/mentor-fa24/Mentor.page.tsx'
+import { MentorPage as MentorSp25Page } from '@/patina/pages/past-programs/mentor-sp25/Mentor.page.tsx'
 import { MentorPage as MentorFa25Page } from '@/patina/pages/mentor-fa25/Mentor.page.tsx'
 import { VolunteerPage } from '@/patina/pages/volunteer/Volunteer.page.tsx'
 import { ScholarshipPage } from '@/patina/pages/scholarship/Scholarship.page.tsx'
