+++ js/src/patina/pages/past-programs/mentorship-sp25/Mentorship.page.tsx
@@ -1,14 +1,12 @@
 import { Space, Text } from '@mantine/core'
-import { About } from '@/patina/pages/mentorship-fa24/about/About.tsx'
-// import { Apply } from '@/patina/pages/mentorship/apply/TimelineSection.tsx'
-import { Gain } from '@/patina/pages/mentorship-fa24/gain/Gain.tsx'
+import { About } from '@/patina/pages/past-programs/mentorship-sp25/about/About.tsx'
+import { Gain } from '@/patina/pages/past-programs/mentorship-sp25/gain/Gain.tsx'
 import { Hero } from '@/patina/components/Hero'
-// import { Details } from '@/patina/pages/mentorship/details/Details.tsx'
-// import { Impact } from '@/patina/pages/mentorship/impact/Impact.tsx'
 import { ContentPage } from '@/patina/components/ContentPage.tsx'
 import { imageUrls } from '@/patina/assets/images'
 import styles from './Mentorship.module.css'
-import { ProgramDetails } from '@/patina/pages/mentorship-fa24/program/ProgramDetails.tsx'
+import { ProgramDetails } from '@/patina/pages/past-programs/mentorship-sp25/program/ProgramDetails.tsx'
+import { Tracks } from '@/patina/pages/past-programs/mentorship-sp25/tracks/Tracks.tsx'
 
 /**
  * MentorshipPage component renders a mentorship page layout with multiple sections.
@@ -22,33 +20,31 @@ export function MentorshipPage() {
           'The Patina Network mentorship program will pair you with an experienced mentor in your field and foster a safe environment to explore how culture, career, and life intersect.'
         }
         <p />
-        {
-          'Sophomores or juniors currently enrolled in a US undergraduate program can apply.'
-        }
-        <br />
-        {'Applications are closed for Fall 2024.'}
+        {'Students enrolled in an undergraduate program based in NYC are eligible. ' +
+          'Freshmen and sophomores may get the most value from the programming but all are encouraged to apply!'}
+        <p />
+        {'Applications for Spring 2025 are now open!'}
       </Text>
     </div>
   )
   return (
     <ContentPage>
       <Hero
-        title="Patina Network Mentorship"
+        title="Patina Mentorship Spring '25"
         details={message}
         img={imageUrls.mentorshipHero.src}
         alt={imageUrls.mentorshipHero.alt}
-        buttonLink="https://docs.google.com/forms/d/195CMB4BcxKAyh72im-eWWCj10SFRB2bFcQGMmHYdzPs/viewform"
-        buttonText="Join the waitlist to be a mentee"
+        buttonLink="https://forms.gle/bFWmWaYBfetfMbfBA"
+        buttonText="Apply to be a mentee!"
         buttonColor={'var(--mantine-color-patina-blue-light)'}
       />
       <About />
       <Space h={64} />
       <ProgramDetails />
+      <Space h={36} />
+      <Tracks />
       <Space h={64} />
       <Gain />
-      {/*<Details />*/}
-      {/*<Apply />*/}
-      {/*<Impact />*/}
     </ContentPage>
   )
 }
